No need of local/remote abstractions
