--- proxy/core/work/local.py
@@ -1,33 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-    proxy.py
-    ~~~~~~~~
-    ⚡⚡⚡ Fast, Lightweight, Pluggable, TLS interception capable proxy server focused on
-    Network monitoring, controls & Application development, testing, debugging.
-
-    :copyright: (c) 2013-present by Abhinav Singh and contributors.
-    :license: BSD, see LICENSE for more details.
-"""
-import asyncio
-import logging
-from typing import Any, Optional
-
-
-logger = logging.getLogger(__name__)
-
-
-class LocalExecutor:
-    """A threadless executor implementation which uses a queue to receive new work."""
-
-    def __init__(self, *args: Any, **kwargs: Any) -> None:
-        super().__init__(*args, **kwargs)
-        self._loop: Optional[asyncio.AbstractEventLoop] = None
-
-    @property
-    def loop(self) -> Optional[asyncio.AbstractEventLoop]:
-        if self._loop is None:
-            self._loop = asyncio.get_event_loop_policy().new_event_loop()
-        return self._loop
-
-    def work_queue_fileno(self) -> Optional[int]:
-        return None
